Sync pgbadger, pod, and README, and fix some syntax errors
pgbadger, doc/pgBadger.pod, and README were out of sync. Fix by running
"make", and slight editing of pgbadger when the pod file was more
correct.

Also, the input was using "\" line-continuations which would need to be
\\ for pod to correctly handle them. Fix by formatting as a single line
instead, which is better suitable for cut-and-paste (and there's already
longer lines in the surrounding text). In plus, crontabs do not support \
anyway. In passing, fix +%F crontab example to be +\%F because % is
magic in crontabs.