Add support to http or ftp remote PostgreSQL log file download. The log file is parsed during the download using curl command and never saved to disk. With ssh remote log parsing you can use uri as comand line argument to specify the PostgreSQL log file.
	ssh://localhost/postgresql-10-main.log
	http://localhost/postgresql-10-main.log.gz
	ftp://localhost/postgresql-10-main.log

with http and ftp protocol you need to specify the log file format
at end of the uri:

	http://localhost/postgresql-10-main.log:stderr

You can specify multiple uri for log files to be parsed. This is
usefull when you have pgbouncer log file on a remote host and
PostgreSQL logs in the local host.

Old syntax to parse remote log file using -r option is still
working but is obsolete and might be removed in two or three
major versions.