Fix remote paths when using SSH.
When specifying a directory instead of a remote file, the program aborts
with a message like:

% pgbadger  ssh://postgres@pghost//postgres/data/log/
FATAL: can't get size of remote file,
       please check what's going wrong with command:
       ssh postgres@pghost "ls -l postgresql-Fri.log"

The problem is that the set_file_list function does not include the path
on the remote host, managing each file path as relative. Therefore
now there is an addition of $dirs before pushing the path to the @lfiles array.
However, this is due only when there is a directory, that is when the number
of remote files on each function invocation is not equal at one.