Implement begin/end hours without the date.
See #495

The idea is to check the '--begin' and '--end' parameters to see if they
represent only a time instead of a timestamp, and then accept them placing the
value into '$from_hour' or '$to_hour' internal variables.
Then, in 'skip_unwanted_lines' check if the time of the line is not between
the limits, similarly to what is done with the timestamp check.
Return the same values as the timestamp check.