Add new configuration option --month-report to be able to build monthly incremental reports.
By default pgBadger in incremental mode only compute daily and weekly reports.
If you want monthly cumulative reports you will have to use a separate command
to specify the report to build. For example to build a report for August 2019:

    pgbadger --month-report 2919-08 /var/www/pg_reports/

this will add a link to the month name into the calendar view of incremental
reports to look at monthly report. The report for a current month can be run
every day it is entirely rebuilt each time. The monthly report is not built by
default because it could take lot of time following the amount of data.

Like with --rebuild option, if you use per database report you must add
the -E option to the command.

Thanks to hvisage and Deyu Tian for the feature request.