Allow to fully separate statistics build and HTML report build in incremental mode without having to read a log file.
For example it is possible to run pgbadger each hours as follow:

    pgbadger -I -O "/out-dir/data" --noreport /var/log/postgresql*.log

thanks to `--noreport` it will just create the data binary files in "/out-dir/data".

Then for example you can make reports each night for the next day
in a separate directory, let's say `/out-dir/reports`:

    pgbadger -I -l "/out-dir/data/LAST_PARSED" -H "/out-dir/reports" /out-dir/data/2020/02/19/*.bin

This require to set the path to the last parsed information, the path
where HTML reports will be written and the binary data file of the day.