Add command --ssh-sudo to run commands over ssh as sudo
This feature comes in handy in environments where you
are expected to log into machines with you own user credentials
but then switch to using service account on that machine. This
feature will only work when no password entry is required for sudo.
Also this feature does not allow the use of wildcard filenames because
of how shell expension works. Also requires sudo to be installed on the
target machine.